Move Plated type-class to monocle.function package
