+++ test/src/test/scala/monocle/std/ListSpec.scala
@@ -1,6 +1,7 @@
 package monocle.std
 
-import monocle.{MonocleSuite, Plated}
+import monocle.MonocleSuite
+import monocle.function.Plated._
 import monocle.law.discipline.{IsoTests, TraversalTests}
 import monocle.law.discipline.function._
 
@@ -15,5 +16,5 @@ class ListSpec extends MonocleSuite {
   checkAll("index List", IndexTests.defaultIntIndex[List[Int], Int])
   checkAll("filterIndex List", FilterIndexTests.evenIndex[List[Int], Int])
 
-  checkAll("plated List", TraversalTests(Plated.plate[List[Int]]))
+  checkAll("plated List", TraversalTests(plate[List[Int]]))
 }
