+++ test/src/test/scala/monocle/std/TreeSpec.scala
@@ -1,6 +1,7 @@
 package monocle.std
 
-import monocle.{MonocleSuite, Plated}
+import monocle.MonocleSuite
+import monocle.function.Plated._
 import monocle.law.discipline.{LensTests, TraversalTests}
 import monocle.law.discipline.function.{EachTests, ReverseTests}
 
@@ -15,5 +16,5 @@ class TreeSpec extends MonocleSuite {
   checkAll("each Tree", EachTests[Tree[Int], Int])
   checkAll("reverse Tree", ReverseTests[Tree[Int], Tree[Int]])
 
-  checkAll("plated Tree", TraversalTests(Plated.plate[Tree[Int]]))
+  checkAll("plated Tree", TraversalTests(plate[Tree[Int]]))
 }
